Repair diskless startup
Without disks, we needed the whole cluster stack to be functional before
reporting that startup was successful.  Unfortunately this wasn't
possible because systemd wouldn't start the rest of the stack until we
reported success.

So now we allow corosync to start in parallel and detatch+succeed once
we detect it.