Fix: sbd-inquisitor: refuse to start if any of the configured device names is invalid
Any device name that is not specified with an absolute path is invalid.
SBD should refuse to start in this case.

Besides, for the case of invalid SBD_DEVICE, it doesn't make sense to
print the usage of the command line. This commit avoids that and only
outputs the error itself.