Fix: sbd-cluster: check quorum-update-time < watchdog-timeout
Checking for this kind of minimum-stall-time probably doesn't make sense
as it will be seen on all nodes involved and thus a delayed suicide
should be uncritical as critical cluster-decisions will be delayed
likewise. More critical is the theoretically undefined time after that
base delay that may pass till new memberships are actually formed
as this may be between 0-inf with the partitions. So the clock should
probably start ticking afterwards and lead to suicide when watchdog-timeout
is exceeded.