Fix: sbd-inquisitor: take the defaults for the options set in sysconfig with empty strings
Prior to 6c1c641, options used to be set in sbd.sysconfig template with
empty strings like:
```
SBD_PACEMAKER=
```

With invocations of sbd.sh by sbd.service, the defaults would be taken
with calls like:
```
: ${SBD_PACEMAKER:="true"}
```

But with that commit, for some of the options, either wrong values are
taken or parsing errors occur. This might break existing setups that get
upgraded with an old sysconfig.